Fix order of some things in install script.
My test run of this script exited with a `0` code right after the line `success 'App built!'`. It did not add anything to /etc/rc.local, and didn't even print out the text that comes right before that. I'm trying to reshuffle the order of commands to see if that works.