Wait for subscript to terminate _before_ exiting
"Wait, you mean it doesn't do this?"
Yeah, this is why all the browsers kept on closing after the xte command. So now I'm storing the PID of startup-subscript and using bash's `wait` keyword to wait for it to close properly. And I threw in a psuedo-DOS `pause` as well.