Switch to yarn for faster start-up
In my testing, yarn seems to be much faster at installing packages than
npm - especially on a raspberry pi. This means faster start-up times, as
we reinstall all of the packages as part of the update process.

The install script and startup script have also been updated to use
yarn. Additionally, the startup script will make sure that it's
installed - for existing slideshows out in the wild.