Wait for server startup betterer
The method detailed in my previous commit worked, but it made Chromium
complain for a bit that it couldn't connect to the server. This would
probably of scared the end user. Now we'll just create a temporary file
from the express server so that Chromium is only started when we are
absolutely certain that the server is running.