Add warning about exFAT USB drives in README
By default, exFAT filesystems are not compatible with Linux. This can be fixed with the `exfat-fuse` and `exfat-utils` packages, but apparently `usbmount` has been unmaintained for quite some time and is absolutely not compatible with exFAT filesystems. So, until we have a replacement for `usbmount` we just need to warn the user about this and point them towards a guide on reformatting their USB's.